Hero Section Complete
The hero section styles are complete. However, the navigation
functionality for the navbar and menu links has not been completed
and will only be so at the end of the project. The functionality for
opening and closing the mobile menu is complete, although it has not
been tested for merging yet (coming next commit).