UI Tests & Test Automatization
UI tests perfomed in the mobile menu are not yet fully implemented.
Currently, they only test the opening and closing functionality,
as the navigation functionality of the links is also not implemented.
The same applies to the desktop navbar links.

Automatization of tests using github actions for commits to main and dev
branches as well as for pull requests on main, has been added.