Merge Hero Section From Dev Branch
* Hero Section Base Phone Viewport Styles

Styles for base phone viewport 385x854 done.
No media queries yet, up next comes media queries
for mobile devices and mobile menu functionality.

* Fix Html Selector Typo

* Mobile Menu Structure & General Changes

General changes to related files and some html structure addition.
Next commit will contain the complete media queries for mobile devices,
and the complete mobile menu functionality tested and ready to be merged
to the main branch.

* Hero Section Complete

The hero section styles are complete. However, the navigation
functionality for the navbar and menu links has not been completed
and will only be so at the end of the project. The functionality for
opening and closing the mobile menu is complete, although it has not
been tested for merging yet (coming next commit).

* UI Tests & Test Automatization

UI tests perfomed in the mobile menu are not yet fully implemented.
Currently, they only test the opening and closing functionality,
as the navigation functionality of the links is also not implemented.
The same applies to the desktop navbar links.

Automatization of tests using github actions for commits to main and dev
branches as well as for pull requests on main, has been added.